Create DAO with working single entity mock implementation
A lot more needs to be done to make this into the fully fledged DAO i
have in mind.  Where there is now only a single entity, there should be
a full schema of about a dozen entities and relationships.  And where
there is now a mock implementation, and stubs of file and emacsql based
implementations, at least the file-based implementation needs to be
fully functional for phase 1.  The emacsql based implementation will
probably also need to be fully implemented for some later phase, notably
when i want to bring other languages into the picture.